Notificaciones y confirmacion de eliminación

+++ src/app/productos/productos.component.ts
@@ -53,7 +53,10 @@ export class ProductosComponent implements OnInit {
 
   public eliminarProducto(i: number) {
     console.log("Eliminando producto..." + this.productos[i].id);
-    this.productoService.eliminarProducto(this.productos[i].id);
+    if(confirm("Desea borrar este producto de la base de datos?")) {
+      console.log("Implement delete functionality here");
+      this.productoService.eliminarProducto(this.productos[i].id);
+    }
   }
 
   togglePanels(index: number) {
